Add files via upload
The above is a c function to calculate the fine for overdue library books based on the number of days a book is overdue. Then used the same to calculate the book ID(integer) ,due date(integer) and return date(integer). Ive also used the if...else statement to determine the fine rate as per specified rates. The C program has displayed ; Book ID, Due date and Return date.