Add files via upload
A C program to read the yearly salary of an employee and finding tax based on the given conditions. The conditions to calculate tax are as follows, if salary is: Less than or equal to 400,000 - Tax will be 1% ■ ii Greater than 400,000 and Less than 1000,000 - Tax will be 1% for 400,000 amount and 15% on remaining amount ■ iii Greater than 1000,000 - Tax will be 1% for 400, 000, 15% for next 600, 000 and 25% on remaining amount.