Add files via upload
A C program to calculate and display the electricity bill for a given customer based on the units consumed. The program considered different charging rates and applied surcharges as specified. The charges are as follow: Units Charges per unit (Ksh) upto 300 175 301 and above but less than 600 200 600 and above but less than 800 300 800 and above 400.